% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_herzog_results.R
\name{get_herzog_results}
\alias{get_herzog_results}
\title{Collect herzog results for specific ids from your designated url}
\usage{
get_herzog_results(url, ids = NULL)
}
\arguments{
\item{ids}{}
}
\description{
Collect herzog results for specific ids from your designated url
}
